backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
python-multipart==0.0.9
numpy==1.26.4
opencv-python-headless==4.9.0.80